+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 * Provide PHP 8.4 in the static binaries
 * Update all PHP vendor
 * Upgrade SPC version to v2.7.5
+* Use a `PhpRunner` service instead of static functions to run PHP files
 
 ## 1.0.0 (2025-10-10)