+++ CHANGELOG.md
@@ -31,6 +31,7 @@
 ### Documentation
 
 * Rework the installation methods documentation
+* Add documentation for arguments/options of compile & repack commands
 
 ## 1.0.0 (2025-10-10)