+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Add a new `--castor-file` option to specify a custom root file when using the `castor` command
 * Add a `slug()` function to slugify strings
 * Display installation method/path in `castor debug`
+* Add support for PHP 8.5 (except for static binaries which are not available yet)
 
 ### Fixes